Added more tests, split tests into three files
